Unit tests (ignore test result if alternative result is met)
- Some tests fail due to what seems to be a bug in POI. The result is
however acceptable (date cells are formatted as "March 03 2017" when it
should be "March 3 2017"). Since I cannot fix this before it is fixed in
POI and the actual result should be OK in virtually all cases, I allow a
second result. Test is reported as SUCCESS if output is correct and as
IGNORED when the actual result matches the alternative.